requirements.txt
RPi.GPIO
blynk-library-python